Clean hf v1 related codes  & rename v2 directroy
