+++ HoaryFox/Component/Geometry/Stb2Brep.cs
@@ -3,31 +3,31 @@ using System.Collections.Generic;
 using System.Drawing;
 using System.Linq;
 using Grasshopper.Kernel;
-using HoaryFox.Member;
+using HoaryFox.Component_v2.Utils.Geometry;
 using HoaryFox.Properties;
 using Rhino;
 using Rhino.DocObjects;
 using Rhino.Geometry;
-using STBReader;
-using STBReader.Member;
+using STBDotNet.v202;
 
-namespace HoaryFox.Component.Geometry
+namespace HoaryFox.Component_v2.Geometry
 {
     public class Stb2Brep : GH_Component
     {
-        private StbData _stbData;
-
-        private readonly List<List<Brep>> _geometryBreps = new List<List<Brep>>();
+        private ST_BRIDGE _stBridge;
+        private readonly List<List<Brep>> _brepList = new List<List<Brep>>();
 
         public Stb2Brep()
-          : base("Stb to Brep", "S2B", "Read ST-Bridge file and display", "HoaryFox", "Geometry")
+          : base("Stb to Brep", "S2B",
+              "Display ST-Bridge model in Brep",
+              "HoaryFox2", "Geometry")
         {
         }
 
         public override void ClearData()
         {
             base.ClearData();
-            _geometryBreps.Clear();
+            _brepList.Clear();
         }
 
         protected override void RegisterInputParams(GH_InputParamManager pManager)
@@ -50,80 +50,71 @@ namespace HoaryFox.Component.Geometry
         protected override void SolveInstance(IGH_DataAccess dataAccess)
         {
             var isBake = false;
-            if (!dataAccess.GetData("Data", ref _stbData)) { return; }
+            if (!dataAccess.GetData("Data", ref _stBridge)) { return; }
             if (!dataAccess.GetData("Bake", ref isBake)) { return; }
-            this.MakeBrep(isBake);
+
+            CreateBrep();
+            if (isBake)
+            {
+                BakeBrep();
+            }
 
             for (var i = 0; i < 7; i++)
             {
-                dataAccess.SetDataList(i, _geometryBreps[i]);
+                dataAccess.SetDataList(i, _brepList[i]);
             }
         }
 
         protected override Bitmap Icon => Resource.Brep;
-        public override Guid ComponentGuid => new Guid("7d2f0c4e-4888-4607-8548-592104f6f06f");
+        public override Guid ComponentGuid => new Guid("B2D5EA7F-E75F-406B-8D22-C267B43C5E72");
 
-        private void MakeBrep(bool isBake)
+        private void CreateBrep()
         {
-            var stbFrames = new List<StbFrame>
-            {
-                _stbData.Columns, _stbData.Girders, _stbData.Posts, _stbData.Beams, _stbData.Braces
-            };
-            var breps = new FrameBreps(_stbData);
-
-            foreach (StbFrame frame in stbFrames)
-            {
-                _geometryBreps.Add(breps.Frame(frame));
-            }
-            _geometryBreps.Add(breps.Slab(_stbData.Slabs));
-            _geometryBreps.Add(breps.Wall(_stbData.Walls));
-
-            if (isBake)
-            {
-                this.BakeBreps(stbFrames);
-            }
+            StbMembers member = _stBridge.StbModel.StbMembers;
+            var brepFromStb = new CreateBrepFromStb(_stBridge.StbModel.StbSections, _stBridge.StbModel.StbNodes, new[] { DocumentTolerance(), DocumentAngleTolerance() });
+            _brepList.Add(brepFromStb.Column(member.StbColumns));
+            _brepList.Add(brepFromStb.Girder(member.StbGirders));
+            _brepList.Add(brepFromStb.Post(member.StbPosts));
+            _brepList.Add(brepFromStb.Beam(member.StbBeams));
+            _brepList.Add(brepFromStb.Brace(member.StbBraces));
+            _brepList.Add(brepFromStb.Slab(member.StbSlabs));
+            _brepList.Add(brepFromStb.Wall(member.StbWalls));
         }
 
-        private void BakeBreps(IEnumerable<StbFrame> stbFrames)
+        private void BakeBrep()
         {
             RhinoDoc activeDoc = RhinoDoc.ActiveDoc;
             var parentLayerNames = new[] { "Column", "Girder", "Post", "Beam", "Brace", "Slab", "Wall" };
             Color[] layerColors = { Color.Red, Color.Green, Color.Aquamarine, Color.LightCoral, Color.MediumPurple, Color.DarkGray, Color.CornflowerBlue };
-            Misc.MakeParentLayers(activeDoc, parentLayerNames, layerColors);
+            GeometryBaker.MakeParentLayers(activeDoc, parentLayerNames, layerColors);
 
-            //TODO: このネストは直す
-            List<List<List<string>>> tagList = stbFrames.Select(stbFrame => Misc.GetTag(_stbData, stbFrame)).ToList();
+            Dictionary<string, string>[][] infoArray = Utils.TagUtils.GetAllSectionInfoArray(_stBridge.StbModel.StbMembers);
 
-            foreach ((List<Brep> frameBreps, int index) in _geometryBreps.Select((frameBrep, index) => (frameBrep, index)))
+            foreach ((List<Brep> breps, int index) in _brepList.Select((frameBrep, index) => (frameBrep, index)))
             {
                 Layer parentLayer = activeDoc.Layers.FindName(parentLayerNames[index]);
-                int parentIndex = parentLayer.Index;
                 Guid parentId = parentLayer.Id;
-
-                foreach ((Brep brep, int bIndex) in frameBreps.Select((brep, bIndex) => (brep, bIndex)))
+                foreach ((Brep brep, int bIndex) in breps.Select((brep, bIndex) => (brep, bIndex)))
                 {
                     var objAttr = new ObjectAttributes();
-                    objAttr.SetUserString("Type", parentLayerNames[index]);
 
-                    if (index < 5)
-                    {
-                        List<List<string>> tags = tagList[index];
-                        List<string> tag = tags[bIndex];
-                        Misc.SetFrameUserString(ref objAttr, tag);
+                    Dictionary<string, string>[] infos = infoArray[index];
+                    Dictionary<string, string> info = infos[bIndex];
 
-                        var layer = new Layer { Name = tag[0], ParentLayerId = parentId, Color = layerColors[index] };
-                        int layerIndex = activeDoc.Layers.Add(layer);
-                        if (layerIndex == -1)
-                        {
-                            layer = activeDoc.Layers.FindName(tag[0]);
-                            layerIndex = layer.Index;
-                        }
-                        objAttr.LayerIndex = layerIndex;
-                    }
-                    else
+                    foreach (KeyValuePair<string, string> pair in info)
                     {
-                        objAttr.LayerIndex = parentIndex;
+                        objAttr.SetUserString(pair.Key, pair.Value);
                     }
+
+                    var layer = new Layer { Name = info["name"], ParentLayerId = parentId, Color = layerColors[index] };
+                    int layerIndex = activeDoc.Layers.Add(layer);
+                    if (layerIndex == -1)
+                    {
+                        layer = activeDoc.Layers.FindName(info["name"]);
+                        layerIndex = layer.Index;
+                    }
+                    objAttr.LayerIndex = layerIndex;
+
                     activeDoc.Objects.AddBrep(brep, objAttr);
                 }
             }
