+++ HoaryFox/Component/IO/StbLoader.cs
@@ -1,42 +1,41 @@
 ﻿using System;
 using Grasshopper.Kernel;
-using STBReader;
+using STBDotNet.v202;
 
-namespace HoaryFox.Component.IO
+namespace HoaryFox.Component_v2.IO
 {
     public class StbLoader : GH_Component
     {
         private string _path;
-        private readonly double _lengthTolerance = DocumentTolerance();
-        private readonly double _angleTolerance = DocumentAngleTolerance();
         public override GH_Exposure Exposure => GH_Exposure.primary;
 
         public StbLoader()
-          : base("Load STB file", "Loader", "Read ST-Bridge file and display", "HoaryFox", "IO")
+          : base("Load STB file", "Loader",
+              "Read ST-Bridge file",
+              "HoaryFox2", "IO")
         {
         }
 
-        protected override void RegisterInputParams(GH_InputParamManager pManager)
+        protected override void RegisterInputParams(GH_Component.GH_InputParamManager pManager)
         {
             pManager.AddTextParameter("path", "path", "input ST-Bridge file path", GH_ParamAccess.item);
         }
 
-        protected override void RegisterOutputParams(GH_OutputParamManager pManager)
+        protected override void RegisterOutputParams(GH_Component.GH_OutputParamManager pManager)
         {
             pManager.AddGenericParameter("Data", "D", "output StbData", GH_ParamAccess.item);
         }
 
         protected override void SolveInstance(IGH_DataAccess dataAccess)
         {
-            // 対象の stb の pathを取得
+            // 対象の stb の path を取得
             if (!dataAccess.GetData("path", ref _path)) { return; }
 
-            var stbData = new StbData(_path, _lengthTolerance, _angleTolerance);
-
+            var stbData = (ST_BRIDGE)STBDotNet.Serialization.Serializer.Deserialize(_path);
             dataAccess.SetData(0, stbData);
         }
 
         protected override System.Drawing.Bitmap Icon => Properties.Resource.LoadStb;
-        public override Guid ComponentGuid => new Guid("B8B7631C-BCAE-4549-95F7-1954D4781D24");
+        public override Guid ComponentGuid => new Guid("C1E1CD82-9AC0-479C-A22F-DB7C44F3C77D");
     }
 }
