+++ HoaryFox/Component/Tag/Name/BraceNameTag.cs
@@ -1,17 +1,75 @@
 ﻿using System;
+using System.Collections.Generic;
 using System.Drawing;
-using HoaryFox.Component.Base;
-using STBReader.Member;
+using Grasshopper.Kernel;
+using HoaryFox.Component_v2.Utils;
+using Rhino.Geometry;
+using STBDotNet.v202;
 
-namespace HoaryFox.Component.Tag.Name
+namespace HoaryFox.Component_v2.Tag.Name
 {
-    public class BraceNameTag : NameTagBase
+    public class BraceNameTag : GH_Component
     {
+        private ST_BRIDGE _stBridge;
+        private int _size;
+
+        private readonly List<string> _frameName = new List<string>();
+        private readonly List<Point3d> _framePos = new List<Point3d>();
+
+        public override bool IsPreviewCapable => true;
         public BraceNameTag()
-          : base(name: "Brace Name Tag", nickname: "BraceTag", description: "Display Brace Name Tag", frameType: FrameType.Brace)
+          : base("Brace Name Tag", "BraceTag",
+              "Display Brace Name Tag",
+              "HoaryFox2", "NameTag")
         {
         }
-        protected override Bitmap Icon => Properties.Resource.BraceName;
-        public override Guid ComponentGuid => new Guid("E566DDCB-4192-40B2-8E96-2083207CC5A8");
+
+        public override void ClearData()
+        {
+            base.ClearData();
+            _frameName.Clear();
+            _framePos.Clear();
+        }
+
+        protected override void RegisterInputParams(GH_InputParamManager pManager)
+        {
+            pManager.AddGenericParameter("Data", "D", "input ST-Bridge file data", GH_ParamAccess.item);
+            pManager.AddIntegerParameter("Size", "S", "Tag size", GH_ParamAccess.item, 12);
+        }
+
+        protected override void RegisterOutputParams(GH_OutputParamManager pManager)
+        {
+            pManager.AddTextParameter("NameTag", "NTag", "output name tag", GH_ParamAccess.list);
+        }
+
+        protected override void SolveInstance(IGH_DataAccess dataAccess)
+        {
+            if (!dataAccess.GetData("Data", ref _stBridge)) { return; }
+            if (!dataAccess.GetData("Size", ref _size)) { return; }
+
+            StbNode[] nodes = _stBridge.StbModel.StbNodes;
+            StbBrace[] braces = _stBridge.StbModel.StbMembers.StbBraces;
+            foreach (StbBrace brace in braces)
+            {
+                _frameName.Add(brace.name);
+
+                string idNodeStart = brace.id_node_start;
+                string idNodeEnd = brace.id_node_end;
+                _framePos.Add(TagUtils.GetTagPosition(idNodeStart, idNodeEnd, nodes));
+            }
+            dataAccess.SetDataList(0, _frameName);
+        }
+
+        public override void DrawViewportWires(IGH_PreviewArgs args)
+        {
+            for (var i = 0; i < _frameName.Count; i++)
+            {
+                args.Display.Draw2dText(_frameName[i], Color.Black, _framePos[i], true, _size);
+            }
+        }
+
+        protected override System.Drawing.Bitmap Icon => Properties.Resource.BraceName;
+
+        public override Guid ComponentGuid => new Guid("21F7885E-6321-4C0D-8974-BC40769AAEAE");
     }
 }
