+++ HoaryFox/Component/Utils/Geometry/CreateLineFromStb.cs
