+++ HoaryFox/Component/Utils/Geometry/GeometryBaker.cs
