+++ HoaryFox/Component/Utils/Geometry/SectionCornerPoints.cs
