--- HoaryFox/Component_v2/Tag/Name/PostNameTag.cs
@@ -1,75 +1,0 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Drawing;
-using Grasshopper.Kernel;
-using HoaryFox.Component_v2.Utils;
-using Rhino.Geometry;
-using STBDotNet.v202;
-
-namespace HoaryFox.Component_v2.Tag.Name
-{
-    public class PostNameTag : GH_Component
-    {
-        private ST_BRIDGE _stBridge;
-        private int _size;
-
-        private readonly List<string> _frameName = new List<string>();
-        private readonly List<Point3d> _framePos = new List<Point3d>();
-
-        public override bool IsPreviewCapable => true;
-        public PostNameTag()
-          : base("Post Name Tag", "PostTag",
-              "Display Post Name Tag",
-              "HoaryFox2", "NameTag")
-        {
-        }
-
-        public override void ClearData()
-        {
-            base.ClearData();
-            _frameName.Clear();
-            _framePos.Clear();
-        }
-
-        protected override void RegisterInputParams(GH_InputParamManager pManager)
-        {
-            pManager.AddGenericParameter("Data", "D", "input ST-Bridge file data", GH_ParamAccess.item);
-            pManager.AddIntegerParameter("Size", "S", "Tag size", GH_ParamAccess.item, 12);
-        }
-
-        protected override void RegisterOutputParams(GH_OutputParamManager pManager)
-        {
-            pManager.AddTextParameter("NameTag", "NTag", "output name tag", GH_ParamAccess.list);
-        }
-
-        protected override void SolveInstance(IGH_DataAccess dataAccess)
-        {
-            if (!dataAccess.GetData("Data", ref _stBridge)) { return; }
-            if (!dataAccess.GetData("Size", ref _size)) { return; }
-
-            StbNode[] nodes = _stBridge.StbModel.StbNodes;
-            StbPost[] posts = _stBridge.StbModel.StbMembers.StbPosts;
-            foreach (StbPost post in posts)
-            {
-                _frameName.Add(post.name);
-
-                string idNodeStart = post.id_node_bottom;
-                string idNodeEnd = post.id_node_top;
-                _framePos.Add(TagUtils.GetTagPosition(idNodeStart, idNodeEnd, nodes));
-            }
-            dataAccess.SetDataList(0, _frameName);
-        }
-
-        public override void DrawViewportWires(IGH_PreviewArgs args)
-        {
-            for (var i = 0; i < _frameName.Count; i++)
-            {
-                args.Display.Draw2dText(_frameName[i], Color.Black, _framePos[i], true, _size);
-            }
-        }
-
-        protected override Bitmap Icon => Properties.Resource.PostName;
-
-        public override Guid ComponentGuid => new Guid("AD517629-4CD1-4109-B071-653D80DC6B70");
-    }
-}
